+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [1.19.4-7.3.1] - 2023-04-13
+
+### Added
+
+- Update GUI inventory if changed while opened.
+- Added pattern book item, you can now transfer patterns between decoration blocks.
+
+### Fixed
+
+- Fixed a bug where the item collector would not use the drill storage inventory.
+
 ## [1.19.4-7.2.2] - 2023-03-15
 
 First release of 1.19.4 fabric version.